improve interface of service
